Fix wrong results returned from the save method
The updated mongodb driver no longer supports returning the `ops` array which
contained the entire document(s) inserted, now it only returns the `_id` field.
This breaking change caused objects passed to the save method to have `id: undefined`
set on them and that messed up consumer code (the panoply codebase).

To keep the behavior as close to the old one I'm assigning the returned id to
the same object that was passed to the `save()` method.

References:
- The old driver result docs: https://mongodb.github.io/node-mongodb-native/3.6/api/Collection.html#~insertOneWriteOpResult
- The new driver result docs: https://mongodb.github.io/node-mongodb-native/5.9/interfaces/InsertOneResult.html